requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-allauth==0.60.1
pillow==10.2.0
django-constance==2.8.0
